+++ src/pages/PostsManagerPage.tsx
@@ -17,7 +17,7 @@ import type {
   PostsResponse,
   UsersResponse,
 } from '../types';
-import { useUpdatePost, useDeletePost } from '../api/posts/usePostsMutations';
+import { useDeletePost } from '../api/posts/usePostsMutations';
 import { usePostsStoreSelector } from '../stores/posts/usePostsStore';
 import { useQueryPosts } from '../api/posts/usePostsQueries';
 import { useQueryUsers } from '../api/users/useUsersQueries';
@@ -46,7 +46,10 @@ import {
   TableRow,
   Textarea,
 } from '../shared/ui';
+
+import { useSelectedPostStore } from '../stores/posts/useSelectedPostStore';
 import PostAddDialog from '../widgets/post/PostAddDialog';
+import PostUpdateDialog from '../widgets/post/PostUpdateDialog';
 
 const PostsManager = () => {
   const navigate = useNavigate();
@@ -54,10 +57,9 @@ const PostsManager = () => {
   const queryParams = new URLSearchParams(location.search);
 
   // 상태 관리
-  const { posts, setPosts, updatePost, deletePost } = usePostsStoreSelector([
+  const { posts, setPosts, deletePost } = usePostsStoreSelector([
     'posts',
     'setPosts',
-    'updatePost',
     'deletePost',
   ]);
   const [total, setTotal] = useState(0);
@@ -68,7 +70,7 @@ const PostsManager = () => {
   const [searchQuery, setSearchQuery] = useState(
     queryParams.get('search') || '',
   );
-  const [selectedPost, setSelectedPost] = useState<Post | null>(null);
+  const { selectedPost, setSelectedPost } = useSelectedPostStore();
   const [sortBy, setSortBy] = useState(queryParams.get('sortBy') || '');
   const [sortOrder, setSortOrder] = useState(
     queryParams.get('sortOrder') || 'asc',
@@ -99,7 +101,6 @@ const PostsManager = () => {
     isLoading: postsLoading,
     error: postsError,
   } = useQueryPosts(limit, skip);
-  const { mutateAsync: mutatePostUpdate } = useUpdatePost();
   const { mutateAsync: mutatePostDelete } = useDeletePost();
 
   const {
@@ -196,20 +197,6 @@ const PostsManager = () => {
       console.error('태그별 게시물 가져오기 오류:', error);
     }
     setLoading(false);
-  };
-
-  // 게시물 업데이트
-  const handleUpdatePost = async () => {
-    try {
-      await mutatePostUpdate(selectedPost!, {
-        onSuccess: (updatedPost) => {
-          updatePost(updatedPost);
-          setShowEditDialog(false);
-        },
-      });
-    } catch (error) {
-      console.error('게시물 업데이트 오류:', error);
-    }
   };
 
   // 게시물 삭제
@@ -655,33 +642,10 @@ const PostsManager = () => {
       <PostAddDialog open={showAddDialog} onOpenChange={setShowAddDialog} />
 
       {/* 게시물 수정 대화상자 */}
-      <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>게시물 수정</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Input
-              placeholder="제목"
-              value={selectedPost?.title || ''}
-              onChange={(e) =>
-                selectedPost &&
-                setSelectedPost({ ...selectedPost, title: e.target.value })
-              }
-            />
-            <Textarea
-              rows={15}
-              placeholder="내용"
-              value={selectedPost?.body || ''}
-              onChange={(e) =>
-                selectedPost &&
-                setSelectedPost({ ...selectedPost, body: e.target.value })
-              }
-            />
-            <Button onClick={handleUpdatePost}>게시물 업데이트</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      <PostUpdateDialog
+        open={showEditDialog}
+        onOpenChange={setShowEditDialog}
+      />
 
       {/* 댓글 추가 대화상자 */}
       <Dialog
